+++ tests/test_evaluation/test_rank_based_metrics.py
@@ -2,102 +2,102 @@
 
 import unittest_templates
 
-import pykeen.evaluation.metrics
+import pykeen.evaluation.ranking_metrics
 from tests import cases
 
 
 class AdjustedArithmeticMeanRankTests(cases.RankBasedMetricTestCase):
     """Tests for adjusted arithmetic mean rank."""
 
-    cls = pykeen.evaluation.metrics.AdjustedArithmeticMeanRank
+    cls = pykeen.evaluation.ranking_metrics.AdjustedArithmeticMeanRank
 
 
 class AdjustedArithmeticMeanRankIndexTests(cases.RankBasedMetricTestCase):
     """Tests for adjusted arithmetic mean rank index."""
 
-    cls = pykeen.evaluation.metrics.AdjustedArithmeticMeanRankIndex
+    cls = pykeen.evaluation.ranking_metrics.AdjustedArithmeticMeanRankIndex
 
 
 class ArithmeticMeanRankTests(cases.RankBasedMetricTestCase):
     """Tests for arithmetic mean rank."""
 
-    cls = pykeen.evaluation.metrics.ArithmeticMeanRank
+    cls = pykeen.evaluation.ranking_metrics.ArithmeticMeanRank
 
 
 class CountTests(cases.RankBasedMetricTestCase):
     """Tests for rank count."""
 
-    cls = pykeen.evaluation.metrics.Count
+    cls = pykeen.evaluation.ranking_metrics.Count
 
 
 class GeometricMeanRankTests(cases.RankBasedMetricTestCase):
     """Tests for geometric mean rank."""
 
-    cls = pykeen.evaluation.metrics.GeometricMeanRank
+    cls = pykeen.evaluation.ranking_metrics.GeometricMeanRank
 
 
 class HarmonicMeanRankTests(cases.RankBasedMetricTestCase):
     """Tests for harmonic mean rank."""
 
-    cls = pykeen.evaluation.metrics.HarmonicMeanRank
+    cls = pykeen.evaluation.ranking_metrics.HarmonicMeanRank
 
 
 class HitsAtKTests(cases.RankBasedMetricTestCase):
     """Tests for Hits at k."""
 
-    cls = pykeen.evaluation.metrics.HitsAtK
+    cls = pykeen.evaluation.ranking_metrics.HitsAtK
 
 
 class InverseArithmeticMeanRankTests(cases.RankBasedMetricTestCase):
     """Tests for inverse arithmetic mean rank."""
 
-    cls = pykeen.evaluation.metrics.InverseArithmeticMeanRank
+    cls = pykeen.evaluation.ranking_metrics.InverseArithmeticMeanRank
 
 
 class InverseMedianRankTests(cases.RankBasedMetricTestCase):
     """Tests for inverse median rank."""
 
-    cls = pykeen.evaluation.metrics.InverseMedianRank
+    cls = pykeen.evaluation.ranking_metrics.InverseMedianRank
 
 
 class InverseGeometricMeanRankTests(cases.RankBasedMetricTestCase):
     """Tests for inverse geometric mean rank."""
 
-    cls = pykeen.evaluation.metrics.InverseGeometricMeanRank
+    cls = pykeen.evaluation.ranking_metrics.InverseGeometricMeanRank
 
 
 class InverseHarmonicMeanRankTests(cases.RankBasedMetricTestCase):
     """Tests for inverse harmonic mean rank."""
 
-    cls = pykeen.evaluation.metrics.InverseHarmonicMeanRank
+    cls = pykeen.evaluation.ranking_metrics.InverseHarmonicMeanRank
 
 
 class MedianAbsoluteDeviationTests(cases.RankBasedMetricTestCase):
     """Tests for MAD."""
 
-    cls = pykeen.evaluation.metrics.MedianAbsoluteDeviation
+    cls = pykeen.evaluation.ranking_metrics.MedianAbsoluteDeviation
 
 
 class MedianRankTests(cases.RankBasedMetricTestCase):
     """Tests for median rank."""
 
-    cls = pykeen.evaluation.metrics.MedianRank
+    cls = pykeen.evaluation.ranking_metrics.MedianRank
 
 
 class StandardDeviationTests(cases.RankBasedMetricTestCase):
     """Tests for rank standard deviation."""
 
-    cls = pykeen.evaluation.metrics.StandardDeviation
+    cls = pykeen.evaluation.ranking_metrics.StandardDeviation
 
 
 class VarianceTests(cases.RankBasedMetricTestCase):
     """Tests for rank variance."""
 
-    cls = pykeen.evaluation.metrics.Variance
+    cls = pykeen.evaluation.ranking_metrics.Variance
 
 
-class RankBasedMetricsTest(unittest_templates.MetaTestCase[pykeen.evaluation.metrics.RankBasedMetric]):
+class RankBasedMetricsTest(unittest_templates.MetaTestCase[pykeen.evaluation.ranking_metrics.RankBasedMetric]):
     """Test for test coverage for rank-based metrics."""
 
-    base_cls = pykeen.evaluation.metrics.RankBasedMetric
+    base_cls = pykeen.evaluation.ranking_metrics.RankBasedMetric
     base_test = cases.RankBasedMetricTestCase
